refactor(beta/skills): reorganize package around runtime-agnostic composition
The composition layer (SkillsToolkit, SkillPlugin) is no longer local-specific
now that it composes *runtimes of any backend, so the local_skills/ package was
a misnomer:

- Promote toolkit.py and plugin.py to the skills/ top level.
- Move the filesystem loader and LocalRuntime into a per-backend
  runtime/local/ subpackage (loader.py, runtime.py), leaving room for a future
  runtime/memory/. The SkillRuntime protocol stays at runtime/protocol.py.
- Drop the local_skills/ package; rename the toolkit's default name to
  skills_toolkit.

Public API (autogen.beta.tools.skills) is unchanged; only internal module
paths move. Renames preserve git history.
