+++ autogen/beta/tools/skills/__init__.py
@@ -2,9 +2,10 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from .local_skills import SkillPlugin, SkillsToolkit
+from .plugin import SkillPlugin
 from .runtime import LocalRuntime
 from .skill_search import SkillSearchToolkit, SkillsClientConfig
+from .toolkit import SkillsToolkit
 
 __all__ = (
     "LocalRuntime",
